+++ marshmallow/__init__.py
@@ -8,9 +8,8 @@ from marshmallow.schema import (
     UnmarshalResult,
 )
 from marshmallow.decorators import pre_dump, post_dump, pre_load, post_load
-from marshmallow.utils import pprint
+from marshmallow.utils import pprint, missing
 from marshmallow.exceptions import MarshallingError, UnmarshallingError, ValidationError
-from marshmallow.marshalling import missing, null
 
 __version__ = '2.0.0a1.dev'
 __author__ = 'Steven Loria'
@@ -30,5 +29,4 @@ __all__ = [
     'UnmarshallingError',
     'ValidationError',
     'missing',
-    'null',
 ]