+++ marshmallow/marshalling.py
@@ -11,6 +11,7 @@ and from primitive types.
 from __future__ import unicode_literals
 
 from marshmallow import base, utils
+from marshmallow.utils import missing
 from marshmallow.compat import text_type, iteritems
 from marshmallow.exceptions import (
     ValidationError,
@@ -19,34 +20,7 @@ from marshmallow.exceptions import (
 __all__ = [
     'Marshaller',
     'Unmarshaller',
-    'null',
-    'missing',
 ]
-
-class _Null(object):
-
-    def __bool__(self):
-        return False
-
-    __nonzero__ = __bool__  # PY2 compat
-
-    def __repr__(self):
-        return '<marshmallow.marshalling.null>'
-
-
-class _Missing(_Null):
-
-    def __repr__(self):
-        return '<marshmallow.marshalling.missing>'
-
-
-# Singleton that represents an empty value.
-null = _Null()
-
-# Singleton value that indicates that a field's value is missing from input
-# dict passed to :meth:`Schema.load`. If the field's value is not required,
-# it's ``default`` value is used.
-missing = _Missing()
 
 
 class ErrorStore(object):
@@ -284,7 +258,7 @@ class Unmarshaller(ErrorStore):
                     )
                 if raw_value is missing and field_obj.load_from:
                     raw_value = data.get(field_obj.load_from, missing)
-                if raw_value is missing and field_obj.missing is not null:
+                if raw_value is missing:
                     _miss = field_obj.missing
                     raw_value = _miss() if callable(_miss) else _miss
                 if raw_value is missing and not field_obj.required: